First working tests
