+++ openapi.yaml
@@ -165,12 +165,14 @@ paths:
             schema:
               $ref: '#/components/schemas/tableStructure'
       responses:
-        '201':
+        '200':
           description: Created
         '400':
           description: Bad request
         '401':
           description: Unauthorized
+        '409':
+          description: Conflict
   /table-structures/{id}:
     get:
       summary: Get a table structure by id
