--- src/test/java/de/uol/pgdoener/th1/api/ConverterApiIT.java
@@ -1,22 +1,0 @@
-package de.uol.pgdoener.th1.api;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.test.web.servlet.MockMvc;
-
-@ExtendWith(SpringExtension.class)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.MOCK)
-@AutoConfigureMockMvc
-public class ConverterApiIT {
-
-    @Autowired
-    private MockMvc mockMvc;
-
-    @Test
-    void previewConvertTableEndpoint() throws Exception {
-  }
-}
