--- src/test/java/de/uol/pgdoener/th1/data/repository/StructureRepositoryIT.java
@@ -1,21 +1,0 @@
-package de.uol.pgdoener.th1.data.repository;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-
-@ExtendWith(SpringExtension.class)
-@DataJpaTest
-class StructureRepositoryIT {
-
-    @Autowired
-    private StructureRepository structureRepository;
-
-    @Test
-    void test(){
-
-    }
-
-}
